old_projects/blogs/project_blog/Leszek-blog/requirements.txt
Django==1.11.7
django-shortcuts==1.5
django-taggit==0.17.1
Markdown==2.6.2
pytz==2018.3
